// WebApplicationPlateforme/ClientApp/src/app/shared/Models/RH/equipement.model.ts
export class Equipement {
  id: number;
  type: string;
  nom: string;
  email: string;
  tel: string;
  transfert: string;
  remarque: string;
  etatdir: string;
  date: string;
  datedir: string;
  iddir: string;
  nomdir: string;
  attribut1: number;
  attribut2: string;
  attribut3: string;
  attribut4: string;
  attribut5: string;
  attribut6: string;
  dateenreg: string;
  userNameCreator: string;
  idUserCreator: string;
}